// src/api/generated/model/exerciseSimpleResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Fitness App API
 * Comprehensive API for fitness application with support for workout plans, exercises, and progress tracking. API provides OAuth2 authentication via Keycloak or JWT tokens.
 */
import type { ReferenceDataResponse } from './referenceDataResponse';

export interface ExerciseSimpleResponse {
  id?: number;
  title?: string;
  thumbnailUrl?: string;
  exerciseCategory?: ReferenceDataResponse;
  primaryMuscles?: ReferenceDataResponse[];
}